+++ ns1/resource_record.go
@@ -20,23 +20,29 @@ import (
 var recordTypeStringEnum = NewStringEnum([]string{
 	"A",
 	"AAAA",
-	"ALIAS",
 	"AFSDB",
+	"ALIAS",
 	"CAA",
+	"CERT",
 	"CNAME",
+	"CSYNC",
+	"DHCID",
 	"DNAME",
 	"DS",
 	"HINFO",
+	"HTTPS",
 	"MX",
 	"NAPTR",
 	"NS",
 	"PTR",
 	"RP",
+	"SMIMEA",
 	"SPF",
 	"SRV",
+	"SVCB",
+	"TLSA",
 	"TXT",
 	"URLFWD",
-	"strings",
 })
 
 func recordResource() *schema.Resource {
@@ -160,6 +166,20 @@ It is suggested to migrate to a regular "answers" block. Using Terraform 0.12+, 
 					},
 				},
 			},
+			"blocked_tags": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
+			"tags": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
 		},
 		Create:   RecordCreate,
 		Read:     RecordRead,
@@ -206,9 +226,25 @@ func recordToResourceData(d *schema.ResourceData, r *dns.Record) error {
 	d.Set("type", r.Type)
 	d.Set("ttl", r.TTL)
 
+	if len(r.Tags) > 0 {
+		terraformTags := make(map[string]interface{}, len(r.Tags))
+		for k, v := range r.Tags {
+			terraformTags[k] = v
+		}
+		d.Set("tags", terraformTags)
+	}
+
+	if len(r.BlockedTags) > 0 {
+		terraformBlockedTags := make([]interface{}, 0)
+		for _, v := range r.BlockedTags {
+			terraformBlockedTags = append(terraformBlockedTags, v)
+		}
+		d.Set("blocked_tags", terraformBlockedTags)
+	}
+
 	d.Set("override_ttl", nil)
-	if r.Type == "ALIAS" && r.Override_TTL != nil {
-		err := d.Set("override_ttl", *r.Override_TTL)
+	if r.Type == "ALIAS" && r.OverrideTTL != nil {
+		err := d.Set("override_ttl", *r.OverrideTTL)
 
 		if err != nil {
 			return fmt.Errorf("[DEBUG] Error setting override_ttl for: %s, error: %#v", r.Domain, err)
@@ -378,7 +414,7 @@ func resourceDataToRecord(r *dns.Record, d *schema.ResourceData) error {
 
 	if r.Type == "ALIAS" {
 		Override_TTL := d.Get("override_ttl").(bool)
-		r.Override_TTL = &Override_TTL
+		r.OverrideTTL = &Override_TTL
 	}
 
 	if v, ok := d.GetOk("link"); ok {
@@ -438,6 +474,20 @@ func resourceDataToRecord(r *dns.Record, d *schema.ResourceData) error {
 			r.Regions[region["name"].(string)] = ns1R
 		}
 	}
+
+	// Even though blocked_tags are not evaluated on GET, dual update logic is currently enforced on POST
+	if _, tagsExist := d.GetOk("tags"); tagsExist == true {
+		if _, blockedTagsExist := d.GetOk("blocked_tags"); blockedTagsExist == false {
+			r.BlockedTags = []string{}
+			log.Println("empty 'blocked tags' key added")
+		}
+	}
+	if _, blockedTagsExist := d.GetOk("blocked_tags"); blockedTagsExist == true {
+		if _, tagsExist := d.GetOk("tags"); tagsExist == false {
+			r.Tags = map[string]string{}
+			log.Println("empty 'tags' key added")
+		}
+	}
 	return nil
 }
 
@@ -455,7 +505,31 @@ func removeEmptyMeta(v map[string]interface{}) {
 // RecordCreate creates DNS record in ns1
 func RecordCreate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*ns1.Client)
-	r := dns.NewRecord(d.Get("zone").(string), d.Get("domain").(string), d.Get("type").(string))
+
+	terraformTags := d.Get("tags").(map[string]interface{})
+	tags := make(map[string]string)
+
+	for key, value := range terraformTags {
+		switch value := value.(type) {
+		case string:
+			tags[key] = value
+		}
+	}
+
+	terraformBlockedTags := d.Get("blocked_tags").([]interface{})
+	blockedTags := make([]string, 0)
+
+	for _, v := range terraformBlockedTags {
+		blockedTags = append(blockedTags, v.(string))
+	}
+
+	r := dns.NewRecord(
+		d.Get("zone").(string),
+		d.Get("domain").(string),
+		d.Get("type").(string),
+		tags,
+		blockedTags,
+	)
 	if err := resourceDataToRecord(r, d); err != nil {
 		return err
 	}
@@ -494,7 +568,31 @@ func RecordDelete(d *schema.ResourceData, meta interface{}) error {
 // RecordUpdate updates the given dns record in ns1
 func RecordUpdate(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*ns1.Client)
-	r := dns.NewRecord(d.Get("zone").(string), d.Get("domain").(string), d.Get("type").(string))
+
+	terraformTags := d.Get("tags").(map[string]interface{})
+	tags := make(map[string]string)
+
+	for key, value := range terraformTags {
+		switch value := value.(type) {
+		case string:
+			tags[key] = value
+		}
+	}
+
+	terraformBlockedTags := d.Get("blocked_tags").([]interface{})
+	blockedTags := make([]string, 0)
+
+	for _, v := range terraformBlockedTags {
+		blockedTags = append(blockedTags, v.(string))
+	}
+
+	r := dns.NewRecord(
+		d.Get("zone").(string),
+		d.Get("domain").(string),
+		d.Get("type").(string),
+		tags,
+		blockedTags,
+	)
 	if err := resourceDataToRecord(r, d); err != nil {
 		return err
 	}
@@ -517,7 +615,7 @@ func recordStateFunc(d *schema.ResourceData, meta interface{}) ([]*schema.Resour
 	return []*schema.ResourceData{d}, nil
 }
 
-// ignores case difference between state and resoruce written in terraform file
+// ignores case difference between state and resource written in terraform file
 // so we can keep data consistency between tf state and ddi since ddi is case insensitive
 func caseSensitivityDiffSuppress(k, old, new string, d *schema.ResourceData) bool {
 	return strings.EqualFold(old, new)